python-server/requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai


# Environment Variable Management
python-dotenv
chromadb
wikipedia

# Machine Learning Utilities
numpy
pypdf
